plugins/ai-coding-java/skills/context7-document-server/requirements.txt
sentence-transformers==2.2.2
PyPDF2==3.0.1
beautifulsoup4==4.12.2
markdown==3.5.1
numpy==1.24.3
PyYAML==6.0.1
python-dotenv==1.0.0